Preventing the error/exception handlers from swallowing errors/exceptions
In our testing we realized that these handlers were notifying bugsnag,
but were also swallowing the exceptions/errors (nothing would show up
in logs, or display on screen when display_errors is On).

In general I think it’s wise to keep the native logging in place (the
user can enable/disable this with PHP config in their environment).